requirements.txt
# Core app
streamlit
groq
python-dotenv
httpx

# DL fake-news classifier
transformers
torch

# RAG (Retrieval-Augmented Generation)
sentence-transformers
faiss-cpu
scikit-learn

# Text processing
numpy
pandas

# Optional (URL text extraction if you later want full article text)
# trafilatura
# beautifulsoup4
